Use PasswordInput in both UserSettingsController and LoginForm
