+++ administration/src/components/user-settings/UserSettingsController.tsx
@@ -1,47 +1,10 @@
-import { Button, Callout, Card, H2, InputGroup, InputGroupProps2, Label } from '@blueprintjs/core'
+import { Button, Callout, Card, H2 } from '@blueprintjs/core'
 import React, { useContext, useState } from 'react'
-import { Tooltip2 } from '@blueprintjs/popover2'
 import { useChangePasswordMutation } from '../../generated/graphql'
 import { ProjectConfigContext } from '../../project-configs/ProjectConfigContext'
 import { AuthContext } from '../../AuthProvider'
 import { useAppToaster } from '../AppToaster'
-
-const ShowPasswordButton = (props: { hidden: boolean; onClick: () => void }) => {
-  return (
-    <Tooltip2
-      content={props.hidden ? 'Passwort anzeigen' : 'Passwort verstecken'}
-      renderTarget={({ isOpen, ref, ...tooltipProps }) => (
-        <Button
-          elementRef={ref}
-          {...tooltipProps}
-          minimal
-          icon={props.hidden ? 'eye-open' : 'eye-off'}
-          onClick={props.onClick}
-        />
-      )}
-    />
-  )
-}
-
-const PasswordInput = ({
-  label,
-  setValue,
-  ...otherProps
-}: InputGroupProps2 & { label: string; setValue: (value: string) => void }) => {
-  const [passwordHidden, setPasswordHidden] = useState(true)
-  return (
-    <Label>
-      {label}
-      <InputGroup
-        placeholder={label}
-        {...otherProps}
-        type={passwordHidden ? 'password' : 'text'}
-        onChange={event => setValue(event.currentTarget.value)}
-        rightElement={<ShowPasswordButton hidden={passwordHidden} onClick={() => setPasswordHidden(!passwordHidden)} />}
-      />
-    </Label>
-  )
-}
+import PasswordInput from "../PasswordInput";
 
 const isUpperCase = (value: string) => {
   return value === value.toUpperCase() && value !== value.toLowerCase()
